fix: rename init to resource_version
