+++ apis/block-storage/skus/v1/sku_types.go
@@ -2,20 +2,9 @@ package v1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	storage "github.com/eu-sovereign-cloud/ecp/apis/block-storage"
 	genv1 "github.com/eu-sovereign-cloud/ecp/apis/generated/types"
-)
-
-// StorageSKUResource is the resource name for storage SKUs
-const StorageSKUResource = "storage-skus"
-
-var (
-	StorageSKUGR  = schema.GroupResource{Group: storage.Group, Resource: StorageSKUResource}
-	StorageSKUGVR = schema.GroupVersionResource{
-		Group: storage.Group, Version: storage.Version, Resource: StorageSKUResource,
-	}
 )
 
 // +kubebuilder:object:root=true
@@ -37,4 +26,8 @@ type StorageSKUList struct {
 	metav1.ListMeta `json:"metadata,omitempty"`
 
 	Items []StorageSKU `json:"items"`
+}
+
+func init() {
+	storage.SchemeBuilder.Register(&StorageSKU{}, &StorageSKUList{})
 }
